Address review feedback: malformed references, path aliasing, meta-tests
- Reject malformed Factory reference shapes (not just blank/whitespace) via
  a V0 @scope/name lexical grammar in NormalizeACPAgentProfile.
- Fix acpAgentProfileStorePath to derive the sidecar filename from the
  config path's own stem, so distinct config paths in the same directory no
  longer alias the same persisted profile.
- Fold the ACP agent profile contract and its tests into the existing
  resolution_contract.go/acp_integrations_test.go root files instead of
  adding a new root-level file, removing the need to touch the ownership
  inventory meta-tests/baseline for this feature slice.
- Remove the dead Document.ACPAgentProfile field, which was cloned/tested
  but never loaded, mapped, or persisted by the document owner (the real
  implementation uses an isolated sidecar file).
- Add multi-path isolation coverage, a second-update reload test, and a
  bidirectional named-settings preservation test (backend scope,
  provider/model defaults, runtime logging/metrics, Worker presets, ACP
  integrations) for the ACP agent profile update/resolve operations.
