+++ internal/ownershipinventory/operator_settings_root_contract.go
@@ -11,8 +11,6 @@ import (
 // Confirmed against docs/internal/projects/packaged-service-structure/operator-settings-root-go-inventory.json
 // (INV-SET-TOPLEVEL).
 var OperatorSettingsThinRootContractFiles = []string{
-	"acp_agent_profile_contract.go",
-	"acp_agent_profile_contract_test.go",
 	"acp_integrations.go",
 	"acp_integrations_test.go",
 	"backend_scope.go",
