+++ pkg/services/operator_settings/document_contract.go
@@ -78,12 +78,11 @@ func sentinelForDocumentFailureKind(kind DocumentFailureKind) error {
 // Document is the detached operator document value peers consume from document
 // operations without importing storage or codec construction ports.
 type Document struct {
-	BackendScopeID  string
-	Defaults        DocumentDefaults
-	Runtime         DocumentRuntimeSettings
-	WorkerPresets   []DocumentWorkerPreset
-	Workers         DocumentWorkerSettings
-	ACPAgentProfile *DocumentACPAgentProfile
+	BackendScopeID string
+	Defaults       DocumentDefaults
+	Runtime        DocumentRuntimeSettings
+	WorkerPresets  []DocumentWorkerPreset
+	Workers        DocumentWorkerSettings
 }
 
 // Clone returns a detached document copy.
@@ -95,10 +94,6 @@ func (document Document) Clone() Document {
 		cloned.WorkerPresets = cloneDocumentWorkerPresets(document.WorkerPresets)
 	}
 	cloned.Workers = document.Workers.Clone()
-	if document.ACPAgentProfile != nil {
-		clonedProfile := document.ACPAgentProfile.Clone()
-		cloned.ACPAgentProfile = &clonedProfile
-	}
 	return cloned
 }
